requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
sqlalchemy==2.0.35
boto3==1.37.10
redis==5.2.1
psycopg[binary]==3.2.3
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.4.2
faiss-cpu==1.12.0
